refactor: recover transaction
