+++ packages/web3-providers/src/Web3/Solana/providers/Phantom.ts
@@ -2,6 +2,7 @@ import { injectedPhantomProvider } from '@masknet/injected-script'
 import {
     PhantomMethodType,
     ProviderType,
+    recoverTransaction,
     serializeTransaction,
     type Transaction,
     type Web3Provider,
@@ -54,26 +55,7 @@ export class SolanaPhantomProvider extends SolanaInjectedWalletProvider {
             },
         })
 
-        console.log('DEBUG: result', result)
-
-        if ('serializeMessage' in transaction && transaction.feePayer) {
-            const args = result.message as SolanaWeb3.MessageArgs
-            const transaction = SolanaWeb3.Transaction.populate(
-                new SolanaWeb3.Message(args),
-                result.signatures.map((x) => encode(x)),
-            )
-            return transaction
-        } else {
-            const args = result.message as SolanaWeb3.MessageV0Args
-            const msg = new SolanaWeb3.MessageV0({
-                ...args,
-                staticAccountKeys: args.staticAccountKeys?.map((x) => new SolanaWeb3.PublicKey(x)) || [],
-                compiledInstructions: args.compiledInstructions || [],
-                addressTableLookups: args.addressTableLookups || [],
-            })
-            const message = SolanaWeb3.VersionedMessage.deserialize(msg.serialize())
-            return new SolanaWeb3.VersionedTransaction(message, result.signatures)
-        }
+        return recoverTransaction(transaction, result.message, result.signatures)
     }
 
     override async signTransactions(transactions: Transaction[]) {
