+++ packages/web3-shared/solana/src/index.ts
@@ -9,3 +9,4 @@ export * from './helpers/address.js'
 export * from './helpers/token.js'
 export * from './helpers/domain.js'
 export * from './helpers/serializeTransaction.js'
+export * from './helpers/recoverTransaction.js'
